move entity models to separate folder
