+++ Prototype Application/Paw5/src/main/java/finki/paw5/model/entities/PersonalProfile.java
@@ -1,4 +1,4 @@
-package finki.paw5.model;
+package finki.paw5.model.entities;
 
 import finki.paw5.model.enumerations.*;
 import jakarta.persistence.Column;
